Rename module to Alignment, add guards as subtopic
