+++ docs/index.rst
@@ -59,6 +59,8 @@ These modules are, in increasing order of complexity:
 
 - `Memory <./modules/memory.html>`_: Memory is the concept of persisting state between calls of a chain/agent. LangChain provides a standard interface for memory, a collection of memory implementations, and examples of chains/agents that use memory.
 
+- `Alignment <./modules/alignment.html>`_: Alignment in the context of LLMs is the process of steering a LLM towards a desired behavior or outcome. Langchain provides a set of tools to aid in this process. Currently, the primary tools for alignment are "guards" which aim to prevent unwanted output from reaching the user and unwanted user input from reaching the LLM. Alignment is an active area of research and designing LLMs which are aligned at output time not after output filtering is an open problem.
+
 
 .. toctree::
    :maxdepth: 1
@@ -72,7 +74,7 @@ These modules are, in increasing order of complexity:
    ./modules/chains.md
    ./modules/agents.md
    ./modules/memory.md
-   ./modules/guards.md
+   ./modules/alignment.md
 
 Use Cases
 ----------
