+++ tests/unit_tests/utilities/test_guards.py
@@ -1,6 +1,6 @@
 import pytest
 
-from langchain.utilities.guards import CustomGuard, RestrictionGuard, StringGuard
+from langchain.alignment.guards import CustomGuard, RestrictionGuard, StringGuard
 from tests.unit_tests.llms.fake_llm import FakeLLM
 from typing import List
